VACMS-20856: Moved again to a new folder
